[Refactor] rename data_elements to structures
