+++ mmaction/datasets/transforms/formatting.py
@@ -3,7 +3,7 @@ import numpy as np
 from mmcv.transforms import BaseTransform, to_tensor
 from mmengine.data import InstanceData, LabelData
 
-from mmaction.data_elements import ActionDataSample
+from mmaction.structures import ActionDataSample
 from mmaction.registry import TRANSFORMS
