+++ mmaction/models/recognizers/base.py
@@ -6,7 +6,7 @@ import torch.nn as nn
 from mmengine.model import BaseModel, merge_dict
 from torch import Tensor
 
-from mmaction.data_elements import ActionDataSample
+from mmaction.structures import ActionDataSample
 from mmaction.registry import MODELS
 from mmaction.utils import (ConfigType, ForwardResults, LabelList,
                             OptConfigType, OptMultiConfig, OptSampleList,
