+++ tests/structures/bbox/test_transforms.py
